in the middle of integrating h2, demo values persisted

+++ src/main/java/de/wartbar/norman/data/Entity.java
@@ -5,9 +5,9 @@ import java.util.*;
 public class Entity {
 
     private static final Object syncObject = new Object();
-    private static int uniqueIdCounter = 0;
+    private static Long uniqueIdCounter = 0L;
 
-    public static int uniqueId() {
+    public static Long uniqueId() {
         synchronized (syncObject) {
             uniqueIdCounter++;
             return uniqueIdCounter;
@@ -17,5 +17,5 @@ public class Entity {
     public List<String> keys = new ArrayList<>();
     public String value = Constants.emptyString;
     public Date date = new Date();
-    public int id = uniqueId();
+    public Long id = uniqueId();
 }
